+++ cli/src/helpers/installPackages.ts
@@ -1,14 +1,15 @@
 import chalk from "chalk";
 import ora from "ora";
+
 import {
   type InstallerOptions,
   type PkgInstallerMap,
 } from "~/installers/index.js";
 import { logger } from "~/utils/logger.js";
 
-type InstallPackagesOptions = {
+type InstallPackagesOptions = InstallerOptions & {
   packages: PkgInstallerMap;
-} & InstallerOptions;
+};
 // This runs the installer for all the packages that the user has selected
 export const installPackages = (options: InstallPackagesOptions) => {
   const { packages } = options;