+++ cli/src/installers/dependencyVersionMap.ts
@@ -4,12 +4,12 @@
  */
 export const dependencyVersionMap = {
   // NextAuth.js
-  "next-auth": "^4.22.1",
-  "@next-auth/prisma-adapter": "^1.0.5",
+  "next-auth": "^4.22.4",
+  "@next-auth/prisma-adapter": "^1.0.7",
 
   // Prisma
-  prisma: "^4.14.0",
-  "@prisma/client": "^4.14.0",
+  prisma: "^5.1.1",
+  "@prisma/client": "^5.1.1",
 
   // Drizzle
   "drizzle-orm": "^0.26.3",
@@ -19,19 +19,18 @@ export const dependencyVersionMap = {
   "@planetscale/database": "^1.7.0",
 
   // TailwindCSS
-  tailwindcss: "^3.3.0",
+  tailwindcss: "^3.3.3",
   autoprefixer: "^10.4.14",
-  postcss: "^8.4.21",
-  prettier: "^2.8.8",
-  "prettier-plugin-tailwindcss": "^0.2.8",
-  "@types/prettier": "^2.7.2",
+  postcss: "^8.4.27",
+  prettier: "^3.0.0",
+  "prettier-plugin-tailwindcss": "^0.5.1",
 
   // tRPC
-  "@trpc/client": "^10.26.0",
-  "@trpc/server": "^10.26.0",
-  "@trpc/react-query": "^10.26.0",
-  "@trpc/next": "^10.26.0",
-  "@tanstack/react-query": "^4.29.7",
-  superjson: "1.12.2",
+  "@trpc/client": "^10.34.0",
+  "@trpc/server": "^10.34.0",
+  "@trpc/react-query": "^10.34.0",
+  "@trpc/next": "^10.34.0",
+  "@tanstack/react-query": "^4.29.25",
+  superjson: "^1.13.1",
 } as const;
 export type AvailableDependencies = keyof typeof dependencyVersionMap;