+++ cli/template/base/src/pages/index.tsx
@@ -1,6 +1,6 @@
-import styles from "./index.module.css";
 import Head from "next/head";
 import Link from "next/link";
+import styles from "./index.module.css";
 
 export default function Home() {
   return (