+++ upgrade/src/app/layout.tsx
@@ -1,8 +1,10 @@
 import "react-diff-view/style/index.css";
 import "~/styles/globals.css";
+
 import { Github } from "lucide-react";
 import { type Metadata } from "next";
 import { Inter } from "next/font/google";
+
 import { cn } from "~/lib/utils";
 
 const inter = Inter({