+++ upgrade/src/components/ui/select.tsx
@@ -3,6 +3,7 @@
 import * as SelectPrimitive from "@radix-ui/react-select";
 import { Check, ChevronDown } from "lucide-react";
 import * as React from "react";
+
 import { cn } from "~/lib/utils";
 
 const Select = SelectPrimitive.Root;